refactor(cache): remove files referencing cache
